Add macOS and Linux support
